requirements.txt
openai==0.28.1
tiktoken
requests
backoff
python-dotenv
